fix: remove Buildkit cache mounts
`dnf` seems to have problems when the cache is used for different architectures. Normally, the cache mounts should speed up the build, but in this case this resides in an error when we try to build for `linux/amd64` **and** `linux/arm64`. Buildkit does not create a cache per platform and mounts the same cache layer instead, which results in wrong index lists and downloaded packages.